+++ assets/release_notes.md
@@ -4,6 +4,9 @@
 ### 0.x.x - January 2022
 ---
 
+- Enable usage of camera flash when scanning barcodes
+- Enable camera toggle when scanning barcodes
+- Updated icon set
 - Removed "upload error report" functionality (instead link to GitHub issues)
 - Updated multiple language translations